--- src/entities/post/index.ts
@@ -1,3 +1,0 @@
-export { Post } from "./ui/Post";
-export { PostActionMenu } from "./ui/PostActionMenu";
-export { PostActionsButtonGroup } from "./ui/PostActionsButtonGroup";
