+++ src/widgets/posts/index.ts
@@ -1,1 +1,4 @@
 export { PostsList } from "./ui/PostsList";
+export { Post } from "./ui/Post";
+export { PostActionMenu } from "./ui/PostActionMenu";
+export { PostActionsButtonGroup } from "./ui/PostActionsButtonGroup";
