+++ src/widgets/posts/ui/Post.tsx
